# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  dto.User:
    properties:
      first_name:
        type: string
      id:
        type: string
      last_name:
        type: string
      login:
        type: string
      middle_name:
        type: string
      password:
        type: string
      phone:
        type: string
      post_code:
        type: string
      role:
        type: string
    type: object
  dto.UserIDRequest:
    properties:
      id:
        type: string
    type: object
  dto.UserRequest:
    properties:
      user:
        $ref: '#/definitions/dto.User'
    type: object
  dto.UserResponse:
    properties:
      user:
        $ref: '#/definitions/dto.User'
    type: object
...
info:
  contact: {}
  description: This is the HLGateway server API.
  title: HLGateway API
  version: "1"
paths:
...
  /api/add-user:
    post:
      consumes:
      - application/json
      description: Add a new user
      parameters:
      - description: Add User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Add user
      tags:
      - UsersService
...
  /api/delete-user:
    delete:
      consumes:
      - application/json
      description: Delete the user by id
      parameters:
      - description: Delete User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserIDRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Delete the user
      tags:
      - UsersService
...
  /api/get-user:
    get:
      consumes:
      - application/json
      description: Get the user by id
      parameters:
      - description: User ID
        in: query
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Get the user
      tags:
      - UsersService
...
  /api/update-user:
    put:
      consumes:
      - application/json
      description: Update user
      parameters:
      - description: Update User Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
      summary: Update user
      tags:
      - UsersService
swagger: "2.0"